requirements.txt
fastapi==0.110.1
uvicorn==0.29.0
catboost==1.2.3
python-multipart==0.0.9
pydantic==2.6.4
Pillow==10.3.0
numpy==1.24.4
